requirements.txt
# Core Dependencies
neo4j==5.27.0
pinecone-client==5.0.1
pyvis==0.3.2
networkx==3.4
tqdm==4.67.1
python-dotenv==1.0.1

# LLM & Embeddings
sentence-transformers==3.3.1  # Local embeddings (all-mpnet-base-v2)
groq==0.11.0                   # Chat completions (openai/gpt-oss-20b)

# Optional: OpenAI (if not using Groq)
openai==1.59.7

# Additional Dependencies
tenacity==9.0.0        # Retry logic for API calls